Ensure agent open files and os file limit display in decimal notation (#16013)
* adding guardrails to ensure agent open files and os file limit display in decimal notation

* changing map[string]float64 to map[string]uint64, which makes large numbers print in regular notation

* fixed float64 conversion, added printf statement to template file to remove sci notation

+++ pkg/util/process_file_stats_linux.go
@@ -34,8 +34,8 @@ func GetProcessFileStats() (*ProcessFileStats, error) {
 	}
 
 	// Retrieving how many file handles the Core Agent process has open and the file limit set on the OS level
-	stats.AgentOpenFiles = float64(rs[process.RLIMIT_NOFILE].Used)
-	stats.OsFileLimit = float64(rs[process.RLIMIT_NOFILE].Soft)
+	stats.AgentOpenFiles = rs[process.RLIMIT_NOFILE].Used
+	stats.OsFileLimit = rs[process.RLIMIT_NOFILE].Soft
 
 	return &stats, nil
 }
